fire_site/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
pillow
requests